Add dayjs library
Added dayjs setup to bootstrap in order to turn on its extensions for parsing format and counting a duration.
Added timeService that counts YouTube timestamps from a string of 'HH:mm:ss' format. These timestamps will be used for fragment preview in an embedded iframe.
Added a time validator that validates string to be in 'HH:mm:ss' format